run.sh: rename script and explicitly run in verbose mode
Rename ci.sh to run.sh.
The main purpose of this script is not to facilitate CI.
It is to give developers who take the course a super-easy and reliable
way to go through it.
We only test it in CI to check that it works when developers use it.

Do the `set -x` explicitly with the `--verbose` key.
We need this setting in CI and when developing the code in the workflow.
Although, developers who take the course don't need all this extra noise
in their console.